pyproject.toml
[tool.poetry]
name = "shareholder-letter-analysis"
version = "0.1.0"
description = ""
authors = ["shaan <skhoslajr@gmail.com>"]
readme = "README.md"
packages = [{include = "shareholder_letter_analysis"}]

[tool.poetry.dependencies]
python = "^3.10"
python-docx = "^0.8.11"
pandas = "^2.0.2"
transformers = "^4.29.2"
torchvision = "^0.15.2"
matplotlib = "^3.7.1"
openai = "^0.27.8"
openai-multi-client = "^0.1.1"
pyarrow = "^12.0.1"
retry = "^0.9.2"
nltk = "^3.8.1"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.2"
ruff = "^0.0.270"
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
